chore: Refactor out reusable csrf-token fetcher
