+++ tests/integration_tests/tasks/test_utils.py
@@ -29,35 +29,36 @@ from tests.integration_tests.test_app import app
     ],
     ids=["Without trailing slash", "With trailing slash"],
 )
-@mock.patch("superset.tasks.cache.fetch_csrf_token")
 @mock.patch("superset.tasks.cache.request.Request")
 @mock.patch("superset.tasks.cache.request.urlopen")
-def test_fetch_url(mock_urlopen, mock_request_cls, mock_fetch_csrf_token, base_url):
-    from superset.tasks.cache import fetch_url
+def test_fetch_csrf_token(mock_urlopen, mock_request_cls, base_url, app_context):
+    from superset.tasks.utils import fetch_csrf_token
 
     mock_request = mock.MagicMock()
     mock_request_cls.return_value = mock_request
 
-    mock_urlopen.return_value = mock.MagicMock()
-    mock_urlopen.return_value.code = 200
+    mock_response = mock.MagicMock()
+    mock_urlopen.return_value.__enter__.return_value = mock_response
 
-    initial_headers = {"Cookie": "cookie", "key": "value"}
-    csrf_headers = initial_headers | {"X-CSRF-Token": "csrf_token"}
-    mock_fetch_csrf_token.return_value = csrf_headers
+    mock_response.status = 200
+    mock_response.read.return_value = b'{"result": "csrf_token"}'
+    mock_response.headers.get_all.return_value = [
+        "session=new_session_cookie",
+        "async-token=websocket_cookie",
+    ]
 
     app.config["WEBDRIVER_BASEURL"] = base_url
-    data = "data"
-    data_encoded = b"data"
+    headers = {"Cookie": "original_session_cookie"}
 
-    result = fetch_url(data, initial_headers)
+    result_headers = fetch_csrf_token(headers)
 
-    assert data == result["success"]
-    mock_fetch_csrf_token.assert_called_once_with(initial_headers)
-    mock_request_cls.assert_called_once_with(
-        "http://base-url/api/v1/chart/warm_up_cache",
-        data=data_encoded,
-        headers=csrf_headers,
-        method="PUT",
+    mock_request_cls.assert_called_with(
+        "http://base-url/api/v1/security/csrf_token/",
+        headers=headers,
+        method="GET",
     )
+
+    assert result_headers["X-CSRF-Token"] == "csrf_token"
+    assert result_headers["Cookie"] == "new_session_cookie"
     # assert the same Request object is used
     mock_urlopen.assert_called_once_with(mock_request, timeout=mock.ANY)
